+++ Nudge/UI/Common/InformationButton.swift
@@ -19,35 +19,37 @@ struct InformationButton: View {
     }
     
     private var informationButton: some View {
-        if appState.deviceSupportedByOSVersion {
-            guard OSVersionRequirementVariables.aboutUpdateURL != "" else { return AnyView(EmptyView()) }
+        guard OSVersionRequirementVariables.aboutUpdateURL != "" else { return AnyView(EmptyView()) }
 
-            return AnyView(
-                Button(action: UIUtilities().openMoreInfo) {
-                    Text(.init(UserInterfaceVariables.informationButtonText.localized(desiredLanguage: getDesiredLanguage(locale: appState.locale))))
-                        .foregroundColor(dynamicTextColor)
-                }
-                    .buttonStyle(.plain)
-                    .help("Click for more information about the security update".localized(desiredLanguage: getDesiredLanguage(locale: appState.locale)))
-                    .onHoverEffect()
-            )
-        } else {
-            guard OSVersionRequirementVariables.unsupportedURL != "" else { return AnyView(EmptyView()) }
-
-            return AnyView(
-                Button(action: UIUtilities().openMoreInfoUnsupported) {
-                    Text(.init(UserInterfaceVariables.informationButtonTextUnsupported.localized(desiredLanguage: getDesiredLanguage(locale: appState.locale))))
-                        .foregroundColor(dynamicTextColor)
-                }
+        return AnyView(
+            Button(action: UIUtilities().openMoreInfo) {
+                Text(.init(UserInterfaceVariables.informationButtonText.localized(desiredLanguage: getDesiredLanguage(locale: appState.locale))))
+                    .foregroundColor(dynamicTextColor)
+            }
                 .buttonStyle(.plain)
-                .help("Click for more information about replacing your device".localized(desiredLanguage: getDesiredLanguage(locale: appState.locale)))
+                .help("Click for more information about the security update.".localized(desiredLanguage: getDesiredLanguage(locale: appState.locale)))
                 .onHoverEffect()
-            )
-        }
+        )
     }
     
     private var dynamicTextColor: Color {
         colorScheme == .light ? Color.accessibleSecondaryLight : Color.accessibleSecondaryDark
+    }
+}
+
+// Technically not information button as this is using the actionButtonTextUnsupported
+struct InformationButtonAsAction: View {
+    @EnvironmentObject var appState: AppState
+    @Environment(\.colorScheme) var colorScheme
+    
+    var body: some View {
+        Button(action: {
+            UIUtilities().openMoreInfo()
+            UIUtilities().postUpdateDeviceActions(userClicked: true, unSupportedUI: true)
+        }) {
+            Text(.init(UserInterfaceVariables.actionButtonTextUnsupported.localized(desiredLanguage: getDesiredLanguage(locale: appState.locale))))
+        }
+        .help("Click for more information about replacing your device.".localized(desiredLanguage: getDesiredLanguage(locale: appState.locale)))
     }
 }